testing: refactored logs directory management
